+++ src/main.rs
@@ -3,73 +3,19 @@ use std::str::FromStr;
 use std::sync::Arc;
 
 use actix::prelude::*;
-use actix_web::{get, middleware, web, App, HttpRequest, HttpResponse, HttpServer, Responder};
-use actix_web_actors::ws;
-use color_eyre::{eyre::format_err, Result};
+use actix_web::{middleware, web, App, HttpServer};
+use color_eyre::eyre::format_err;
 use objects::{LineReference, StopReference};
 use tracing_subscriber::{filter::targets::Targets, layer::SubscriberExt, util::SubscriberInitExt};
 
+mod actions;
 mod actors;
 mod messages;
 mod objects;
 mod status;
 mod templates;
+use actions::*;
 use actors::*;
-
-#[get("/ws")]
-async fn websocket(
-    req: HttpRequest,
-    stream: web::Payload,
-    central: web::Data<Addr<CentralDispatch>>,
-) -> Result<HttpResponse, actix_web::Error> {
-    tracing::info!("new websocket");
-
-    ws::start(
-        SessionActor {
-            central: central.as_ref().clone(),
-            watching: Watching::Index,
-        },
-        &req,
-        stream,
-    )
-}
-
-#[get("/ws/lines/{line_ref}")]
-async fn line_websocket(
-    req: HttpRequest,
-    line_ref: web::Path<String>,
-    stream: web::Payload,
-    central: web::Data<Addr<CentralDispatch>>,
-) -> Result<HttpResponse, actix_web::Error> {
-    tracing::info!("new websocket watching {line_ref}");
-
-    ws::start(
-        SessionActor {
-            central: central.as_ref().clone(),
-            watching: Watching::Line(line_ref.clone()),
-        },
-        &req,
-        stream,
-    )
-}
-
-#[get("/")]
-async fn index() -> impl Responder {
-    let s = templates::TEMPLATES
-        .render("index.html", &tera::Context::new())
-        .unwrap();
-    HttpResponse::Ok().content_type("text/html").body(s)
-}
-
-#[get("/lines/{id}")]
-async fn line(line_ref: web::Path<String>) -> impl Responder {
-    let mut context = tera::Context::new();
-    context.insert("line_ref", &line_ref.as_str());
-    let s = templates::TEMPLATES
-        .render("line_index.html", &context)
-        .unwrap();
-    HttpResponse::Ok().content_type("text/html").body(s)
-}
 
 fn setup_logger() {
     color_eyre::install().unwrap();
